Move instance constructor samples
