+++ docs/csharp/programming-guide/classes-and-structs/snippets/instance-constructors/coords/Program.cs
@@ -1,0 +1,70 @@
+﻿using System;
+
+namespace CoordsWithParameterlessConstructorOnly
+{
+    //<Snippet1>
+    class Coords
+    {
+        public int x, y;
+    
+        // constructor
+        public Coords()
+        {
+            x = 0;
+            y = 0;
+        }
+    }
+    //</Snippet1>
+}
+
+namespace CoordsWithTwoArgumentsConstructorAndToString
+{
+    //<Snippet4>
+    class Coords
+    {
+        public int x, y;
+
+        // Default constructor.
+        public Coords()
+        {
+            x = 0;
+            y = 0;
+        }
+
+        //<Snippet2>
+        // A constructor with two arguments.
+        public Coords(int x, int y)
+        {
+            this.x = x;
+            this.y = y;
+        }
+        //</Snippet2>
+
+        // Override the ToString method.
+        public override string ToString()
+        {
+            return $"({x},{y})";
+        }
+    }
+
+    class MainClass
+    {
+        static void Main()
+        {
+            //<Snippet3>
+            var p1 = new Coords();
+            var p2 = new Coords(5, 3);
+            //</Snippet3>
+
+            // Display the results using the overriden ToString method.
+            Console.WriteLine($"Coords #1 at {p1}");
+            Console.WriteLine($"Coords #2 at {p2}");
+            Console.ReadKey();
+        }
+    }
+    /* Output:
+     Coords #1 at (0,0)
+     Coords #2 at (5,3)
+    */
+    //</Snippet4>
+}
