+++ docs/csharp/programming-guide/classes-and-structs/snippets/instance-constructors/person/Person.csproj
@@ -1,0 +1,9 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>net5.0</TargetFramework>
+    <RootNamespace>Person</RootNamespace>
+  </PropertyGroup>
+
+</Project>
