+++ docs/csharp/programming-guide/classes-and-structs/snippets/instance-constructors/person/Program.cs
@@ -1,0 +1,21 @@
+﻿using System;
+
+public class Person
+{
+    public int age;
+    public string name;
+}
+
+class TestPerson
+{
+    static void Main()
+    {
+        var person = new Person();
+
+        Console.WriteLine($"Name: {person.name}, Age: {person.age}");
+        // Keep the console window open in debug mode.
+        Console.WriteLine("Press any key to exit.");
+        Console.ReadKey();
+    }
+}
+// Output:  Name: , Age: 0
