+++ docs/csharp/programming-guide/classes-and-structs/snippets/instance-constructors/shapes/Shapes.csproj
@@ -1,0 +1,9 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>net5.0</TargetFramework>
+    <RootNamespace>Shapes</RootNamespace>
+  </PropertyGroup>
+
+</Project>
